Move + rename delegate wrappers and GMA extensions into Tracking/
Move every adapter file into the Tracking/ subfolder and drop the RC*/
RCAdMob* prefixes by nesting the types inside the Tracking namespace.

File moves (git mv preserves history):
- RCAdMobBannerViewDelegate.swift          -> Tracking/BannerViewDelegate.swift
- RCAdMobFullScreenContentDelegate.swift   -> Tracking/FullScreenContentDelegate.swift
- RCAdMobNativeAdDelegate.swift            -> Tracking/NativeAdDelegate.swift
- RCAdMobNativeAdLoader.swift              -> Tracking/AdLoader+Tracking.swift
- BannerView+RCAdMob.swift                 -> Tracking/BannerView+Tracking.swift
- FullScreenAd+RCAdMob.swift               -> Tracking/FullScreenAd+Tracking.swift
- FullScreenAd+RCAdMob+Completion.swift    -> Tracking/FullScreenAd+Tracking+Completion.swift

Type renames (all internal):
- RCAdMobBannerViewDelegate         -> Tracking.BannerViewDelegate
- RCAdMobFullScreenContentDelegate  -> Tracking.FullScreenContentDelegate
- RCAdMobNativeAdDelegate           -> Tracking.NativeAdDelegate
- RCNativeAdLoaderDelegateProxy     -> Tracking.NativeAdLoaderDelegateProxy
  (kept "Proxy" suffix to avoid clashing with GoogleMobileAds.NativeAdLoaderDelegate)
- RCFullScreenAdTracking (protocol) -> Tracking.FullScreenAd
- FullScreenLoadContext (struct)    -> Tracking.FullScreenLoadContext
- rcSetTrackingFullScreenContentDelegate (helper on the protocol)
                                    -> setTrackingDelegate on Tracking.FullScreenAd
- private file-scoped key enums lose RC prefix
  (RCBannerAssociatedKeys -> BannerAssociatedKeys,
   RCNativeAdLoaderAssociatedKeys -> NativeAdLoaderAssociatedKeys)

Public API surface is unchanged: setTrackingFullScreenContentDelegate(_:),
loadAndTrack(...), and present(from:placement:...) keep their signatures and
behavior; only the internal helpers they forward to were renamed.

Test files still use their RCAdMob* filenames (deferred to commit 6); only
the type references inside them were updated.

Verified: swift build, xcodebuild test on iOS Simulator (87 tests passing),
swiftlint clean.
